Add instructions for using Docker
This is a substantial rewrite of this repo. Instead of using systemd to
create and manage servers, it enables you to run servers via Docker. It
supports multiple containers on the same system, and always grabs the
latest version of server.jar after an update was released.

The old instructions are now in the `advanced` directory.

Signed-off-by: Brian Warner <brian@bdwarner.com>